[BUGFIX] Don't assume pages.no_index exists
The pages.no_index field appears to be added by an
extension which is not and should not be declared
as dependency for EXT:dashboard.

A check if the TCA for pages.no_index exists before
using it as constraint solves the problem while
preserving the behavior if the field is defined.