[FEATURE] Add content changes widget
Adds a new widget which shows:

* Pending publishing state change, if any. Lists records
  which have starttime / endtime in the future.
* Changed records along with name of user who made
  the change.

Attempts to select a total of $limit (default 5) records
while distributing the number of records equally between
the "pending" and "changed" sets, giving preference to
the "pending" set:

* If limit is 5 and 3 "pending" were selected, only 2 "changed"
  will be selected.
* If limit is 5 and no "pending" were selected, all 5 are
  selected from the "changed" set.

If there are no records in a group the table containing
the record data is not shown. if there are no results in
either of the two groups, a "nothing to show" message
is shown instead.